refactor: fold the review-gate findings into the read and details paths
/simplify: collapse the details bodies onto shared `NodeRows`/`StateRows`
primitives so a file and a folder no longer restate the same four rows, drop
the now-unused `StateRows.tsx`, and stop exporting the preview byte budget
that nothing outside the hook reads.

Altitude: `streamTicket` moves out of the download hook into `lib/` so the
preview hook no longer depends on a sibling hook, and the text editor now
names a non-text load as a refusal instead of rendering an empty dialog.

/security-review found nothing to fold in.
