+++ apps/web/src/components/file-browser/details/FileDetails.tsx
@@ -1,21 +1,18 @@
 import type { ListingRow } from '../../../vault/listing';
-import { CopyableValue, DetailRow, DetailSection, DimValue, UNKNOWN } from './DetailsPrimitives';
-import { StateRows } from './StateRows';
+import {
+  DetailRow,
+  DetailSection,
+  DimValue,
+  NodeRows,
+  StateRows,
+  UNKNOWN,
+} from './DetailsPrimitives';
 
 /** One file's current version, as the engine snapshot reports it. */
 export function FileDetails({ row }: { row: ListingRow }) {
   return (
     <dl className="details-list" data-testid="file-details">
-      <DetailSection label="node" />
-      <DetailRow label="name">
-        <CopyableValue value={row.name} label="name" />
-      </DetailRow>
-      <DetailRow label="type">
-        <span className="details-badge">{row.icon}</span>
-      </DetailRow>
-      <DetailRow label="node id">
-        <CopyableValue value={row.key} label="node id" />
-      </DetailRow>
+      <NodeRows row={row} />
 
       <DetailSection label="content" />
       <DetailRow label="size">
@@ -24,7 +21,6 @@ export function FileDetails({ row }: { row: ListingRow }) {
       <DetailRow label="bytes">
         {row.bytes === null ? <DimValue>{UNKNOWN}</DimValue> : row.bytes.toString()}
       </DetailRow>
-      <DetailRow label="modified">{row.modified}</DetailRow>
 
       <StateRows row={row} />
     </dl>
